Move Reykjavik.d.ts under src/Reykjavik
